+++ client/app/pages/queries/utils/updateQuery.jsx
